Added number spinner as a widget
